refactor: convert OG image generation to lazy plugin
Replace build-time PNG generation with on-demand rendering via
Revolution plugin. PNGs are now generated when requested and cached
using the Web Cache API.

Changes:
- Add www/plugins/og-image.ts - intercepts /blog/**/*.png requests
- Use @effectionx/fs for Effection-native file operations
- Remove www/scripts/generate-og-images.ts build script
- Remove generate-og-images task from deno.json
- Remove CI step for pre-generating images

Benefits:
- Server IS the build (cowboyd's principle)
- Same behavior locally and in production
- Issues caught during dev before CI runs
- No separate build step needed
