+++ www/deno.json
@@ -3,8 +3,7 @@
     "dev": "deno run -A @effectionx/watch deno run -A main.tsx",
     "staticalize": "deno run -A jsr:@frontside/staticalize@0.2.2/cli --site http://localhost:8000 --output=built --base=http://localhost:8000",
     "pagefind": "npx pagefind --site built",
-    "test": "deno test --allow-run --allow-write --allow-read",
-    "generate-og-images": "deno run -A scripts/generate-og-images.ts"
+    "test": "deno test --allow-run --allow-write --allow-read"
   },
   "lint": {
     "exclude": [
@@ -52,6 +51,7 @@
     "@types/hast": "npm:@types/hast@3.0.4",
     "effection": "npm:effection@^3.6.1",
     "@effectionx/deno-deploy": "jsr:@effectionx/deno-deploy@^0.3.1",
+    "@effectionx/fs": "npm:@effectionx/fs@^0.2.1",
     "@effectionx/process": "jsr:@effectionx/process@^0.6.2",
     "hast": "npm:hast@^1.0.0",
     "hast-util-from-html": "npm:hast-util-from-html@2.0.3",
